Added all prototype blocks

+++ Robot-Game/RobotGame3D/Assets/Scripts/Scripting/Compiled/Instructions/RotaryActuatorSetSpeedInstruction.cs
@@ -12,7 +12,7 @@ namespace Assets.Scripts.Scripting.Compiled.Instructions {
 		}
 
 		public override async Task Execute(InstructionExecutionArgs args) {
-			int actuatorID = (int)await this.args[0].Execute(args);
+			int actuatorID = (int)(float)await this.args[0].Execute(args);
 			float setpoint = (float)await this.args[1].Execute(args);
 
 			args.Robot.rotaryActuatorManagers[actuatorID].SetActuatorSpeed(setpoint);
